requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
multidict==6.0.2
nextcord==2.0.0a3
numpy==1.23.0
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.26.0
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
Unidecode==1.3.2
urllib3==1.26.9
wsproto==1.1.0
yarl==1.7.2
